Fix incorrect assessment overview memoization (#3304)
* Move assessment overview card to separate file

Also refactor to React.FC.

* Refactor interact button and move to separate file

* Fix deprecations

* Remove unused index param

* Fix incorrect memoization

* Fix deprecations

* Update snapshots
